refactor: view의 위치를 변경
