+++ src/main/java/chess/console/ChessController.java
@@ -4,9 +4,9 @@ import chess.domain.board.Board;
 import chess.domain.command.*;
 import chess.domain.game.ChessGame;
 import chess.domain.piece.PieceFactory;
-import chess.view.InputView;
-import chess.view.OutputView;
-import chess.view.dto.BoardDto;
+import chess.console.view.InputView;
+import chess.console.view.OutputView;
+import chess.console.view.dto.BoardDto;
 
 import java.util.Arrays;
